restructured the MCRHTTPClient classes
