+++ mycore-base/src/main/resources/config/mycore.properties
@@ -73,6 +73,9 @@ MCR.Filter.UserAgent.AcceptInvalid=false
 
   MCR.URIResolver.CachingResolver.Capacity=100
   MCR.URIResolver.CachingResolver.MaxAge=3600000
+  
+ # The HTTP Client in use (currently only by MCR.URIResolver.MCRRESTResolver)
+ MCR.HTTPClient.Class=org.mycore.common.MCRDefaultHTTPClient
 
 ##############################################################################
 # Classes for the commandline interface
